requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
Pillow==11.2.1
qrcode==7.4.2
python-barcode==0.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
drf-nested-routers==0.94.0 